+++ pkg/commands/serviceversion/stage.go
@@ -25,8 +25,7 @@ type StageCommand struct {
 func NewStageCommand(parent argparser.Registerer, g *global.Data) *StageCommand {
 	var c StageCommand
 	c.Globals = g
-	// FIXME: unhide this command when appropriate
-	c.CmdClause = parent.Command("stage", "Stage a Fastly service version").Hidden()
+	c.CmdClause = parent.Command("stage", "Stage a Fastly service version")
 	c.RegisterFlag(argparser.StringFlagOpts{
 		Name:        argparser.FlagServiceIDName,
 		Description: argparser.FlagServiceIDDesc,