+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## [Unreleased]
+- feat(domains): add `description` to `domainv1` endpoints ([#1483](https://github.com/fastly/cli/pull/1483))
 
 ### Breaking:
 
@@ -12,6 +13,7 @@
 
 ### Bug fixes:
 - fix(sso): Don't display the token after authentication. ([#1490](https://github.com/fastly/cli/pull/1490))
+- fix(service-version): Stop hiding the 'stage' and 'unstage' commands. ([#1492](https://github.com/fastly/cli/pull/1492))
 
 ### Dependencies:
 - build(deps): `github.com/fastly/go-fastly/v10` from 10.0.1 to 10.1.0 ([#1476](https://github.com/fastly/cli/pull/1476))